refactor: Merge moe model factory into general model factory
