+++ src/modalities/models/moe/__init__.py
@@ -1,10 +1,8 @@
 from modalities.models.moe.loss_functions import MoECrossEntropyLoss
-from modalities.models.moe.model_factory import get_ep_wrapped_model
 from modalities.models.moe.qwen_model import QwenModel, QwenModelConfig
 
 __all__ = [
     "MoECrossEntropyLoss",
     "QwenModel",
     "QwenModelConfig",
-    "get_ep_wrapped_model",
 ]
